requirements.txt
opencv-python==4.10.0.84
python-dotenv==1.0.1
websockets==14.1
pycairo==1.27.0
PyGObject==3.50.0